Turn off monitoring

+++ src/rastervision/ml_backends/tf_object_detection_api.py
@@ -190,7 +190,8 @@ def terminate_at_exit(process):
     atexit.register(terminate)
 
 
-def train(config_path, output_dir, train_py=None, eval_py=None):
+def train(config_path, output_dir, train_py=None, eval_py=None,
+          do_monitoring=True):
     output_train_dir = join(output_dir, 'train')
     output_eval_dir = join(output_dir, 'eval')
 
@@ -204,21 +205,22 @@ def train(config_path, output_dir, train_py=None, eval_py=None):
     ])
     terminate_at_exit(train_process)
 
-    eval_process = Popen([
-        'python', eval_py, '--logtostderr',
-        '--pipeline_config_path={}'.format(config_path),
-        '--checkpoint_dir={}'.format(output_train_dir),
-        '--eval_dir={}'.format(output_eval_dir)
-    ])
-    terminate_at_exit(eval_process)
+    if do_monitoring:
+        eval_process = Popen([
+            'python', eval_py,
+            '--logtostderr', '--pipeline_config_path={}'.format(config_path),
+            '--checkpoint_dir={}'.format(output_train_dir),
+            '--eval_dir={}'.format(output_eval_dir)])
+        terminate_at_exit(eval_process)
 
-    tensorboard_process = Popen(
-        ['tensorboard', '--logdir={}'.format(output_dir)])
-    terminate_at_exit(tensorboard_process)
+        tensorboard_process = Popen([
+            'tensorboard', '--logdir={}'.format(output_dir)])
+        terminate_at_exit(tensorboard_process)
 
     train_process.wait()
-    eval_process.terminate()
-    tensorboard_process.terminate()
+    if do_monitoring:
+        eval_process.terminate()
+        tensorboard_process.terminate()
 
 
 def get_last_checkpoint_path(train_root_dir):
@@ -597,11 +599,10 @@ class TFObjectDetectionAPI(MLBackend):
             export_py = options.object_detection_options.export_py
 
             # Train model and sync output periodically.
-            start_sync(
-                output_dir,
-                options.output_uri,
-                sync_interval=options.sync_interval)
-            train(config_path, output_dir, train_py=train_py, eval_py=eval_py)
+            start_sync(output_dir, options.output_uri,
+                       sync_interval=options.sync_interval)
+            train(config_path, output_dir, train_py=train_py, eval_py=eval_py,
+                  do_monitoring=options.do_monitoring)
 
             export_inference_graph(
                 output_dir, config_path, output_dir, export_py=export_py)
